Clean messed up files, removing unused import, refactoring error messages
