--- src/main/java/com/nivlalulu/nnpro/dao/InvoiceRepository.java
@@ -1,19 +1,0 @@
-package com.nivlalulu.nnpro.dao;
-
-import com.nivlalulu.nnpro.model.Invoice;
-import com.nivlalulu.nnpro.model.InvoiceItem;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface InvoiceRepository extends JpaRepository<Invoice, UUID> {
-
-    Optional<Invoice> findById(UUID id);
-
-    List<Invoice> findAllByProductListContains(InvoiceItem invoiceItem);
-
-}
