--- src/main/java/com/nivlalulu/nnpro/dto/LoginDto.java
@@ -1,11 +1,0 @@
-package com.nivlalulu.nnpro.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginDto {
-
-
-}
