--- src/main/java/com/nivlalulu/nnpro/dto/RegistrationDto.java
@@ -1,21 +1,0 @@
-package com.nivlalulu.nnpro.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class RegistrationDto {
-
-    private String username;
-
-    private String firstName;
-
-    private String lastName;
-
-    private String email;
-
-    private String password;
-}
