--- src/main/java/com/nivlalulu/nnpro/service/ProductService.java
@@ -1,88 +1,0 @@
-package com.nivlalulu.nnpro.service;
-
-import com.nivlalulu.nnpro.dao.InvoiceRepository;
-import com.nivlalulu.nnpro.dao.ProductRepository;
-import com.nivlalulu.nnpro.dto.InvoiceItemDto;
-import com.nivlalulu.nnpro.model.Invoice;
-import com.nivlalulu.nnpro.model.InvoiceItem;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
-import java.util.UUID;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-public class ProductService {
-
-    private final ProductRepository productRepository;
-
-    private final InvoiceRepository invoiceRepository;
-
-    public InvoiceItemDto createProduct(InvoiceItemDto invoiceItemDto) {
-        Optional<InvoiceItem> isProductExisting = productRepository.findProductByNameAndPrice(invoiceItemDto.getName(), invoiceItemDto.getPrice());
-        if (isProductExisting.isPresent()) {
-            return duplicityCheck(isProductExisting.get(), invoiceItemDto) ?
-                    MappingService.convertToDto(isProductExisting.get()) :
-                    MappingService.convertToDto(productRepository.save(isProductExisting.get()));
-        }
-
-        InvoiceItem invoiceItem = new InvoiceItem(invoiceItemDto.getName(), invoiceItemDto.getQuantity(), invoiceItemDto.getPrice(),
-                invoiceItemDto.getTax(), invoiceItemDto.getTotal());
-        return MappingService.convertToDto(productRepository.save(invoiceItem));
-    }
-
-    public InvoiceItemDto updateProduct(InvoiceItemDto invoiceItemDto) {
-        InvoiceItem invoiceItem = productRepository.findById(invoiceItemDto.getId())
-                .orElseThrow(() -> new RuntimeException(String.format("Product with id %s doesn't exist, can't be updated",
-                        invoiceItemDto.getId())));
-
-        invoiceItem.setName(invoiceItemDto.getName());
-        invoiceItem.setUnitPrice(invoiceItemDto.getPrice());
-        invoiceItem.setTaxPrice(invoiceItemDto.getTax());
-        invoiceItem.setTotalPrice(invoiceItemDto.getTotal());
-        invoiceItem.setQuantity(invoiceItemDto.getQuantity());
-
-        return MappingService.convertToDto(productRepository.save(invoiceItem));
-    }
-
-    public InvoiceItemDto deleteProduct(UUID id) {
-        InvoiceItem invoiceItem = productRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException(String.format("Product with id %s doesn't exist, can't be deleted", id)));
-
-        List<Invoice> listWhichContainsProduct = invoiceRepository.findAllByProductListContains(invoiceItem);
-
-        if (!listWhichContainsProduct.isEmpty()) {
-            throw new RuntimeException(String.format("Product with id %s can't be deleted, is in invoices", id));
-        }
-
-        productRepository.delete(invoiceItem);
-        return MappingService.convertToDto(invoiceItem);
-    }
-
-    public List<InvoiceItemDto> findAllByPrice(BigDecimal price) {
-        return productRepository.findAllByPrice(price).stream().map(MappingService::convertToDto).collect(Collectors.toList());
-    }
-
-    public List<InvoiceItemDto> findAllByNameContaining(String name) {
-        return productRepository.findAllByNameContaining(name).stream().map(MappingService::convertToDto).collect(Collectors.toList());
-    }
-
-    public List<InvoiceItem> findProductsByIds(List<UUID> ids) {
-        return productRepository.findByIdIn(ids);
-    }
-
-    public Optional<InvoiceItem> findProductById(UUID id) {
-        return productRepository.findById(id);
-    }
-
-    protected boolean duplicityCheck(InvoiceItem invoiceItem, InvoiceItemDto invoiceItemDto) {
-        boolean isNameMatching = invoiceItem.getName().equals(invoiceItemDto.getName());
-        boolean isNamePrice = invoiceItem.getUnitPrice().equals(invoiceItemDto.getPrice());
-        return isNameMatching && isNamePrice;
-    }
-}
